+++ src/main/java/dev/the_fireplace/caterpillar/block/DecorationBlock.java
@@ -48,21 +48,18 @@ public class DecorationBlock extends DrillBaseBlock {
     private static final Map<Direction, VoxelShape> SHAPES_RIGHT = new EnumMap<>(Direction.class);
 
     private static final VoxelShape SHAPE_LEFT = Stream.of(
-            Block.box(0, 0, 0, 6, 16, 16),
-            Block.box(6, 6, 6, 16, 10, 10)
+        Block.box(0, 0, 0, 6, 16, 16),
+        Block.box(6, 6, 6, 16, 10, 10)
     ).reduce((v1, v2) -> Shapes.join(v1, v2, BooleanOp.OR)).get();
 
     private static final VoxelShape SHAPE_BASE = Stream.of(
-            Block.box(6, 0, 0, 10, 6, 16),
-            Block.box(10, 0, 0, 16, 16, 16),
-            Block.box(0, 0, 0, 6, 16, 16),
-            Block.box(6, 10, 0, 10, 16, 16),
-            Block.box(6, 6, 16, 10, 10, 31)
+        Block.box(0, 0, 0, 16, 16, 16),
+        Block.box(6, 6, 16, 10, 10, 32)
     ).reduce((v1, v2) -> Shapes.join(v1, v2, BooleanOp.OR)).get();
 
     private static final VoxelShape SHAPE_RIGHT = Stream.of(
-            Block.box(0, 6, 6, 10, 10, 10),
-            Block.box(10, 0, 0, 16, 16, 16)
+        Block.box(0, 6, 6, 10, 10, 10),
+        Block.box(10, 0, 0, 16, 16, 16)
     ).reduce((v1, v2) -> Shapes.join(v1, v2, BooleanOp.OR)).get();
 
     public DecorationBlock(Properties properties) {