+++ src/main/java/dev/the_fireplace/caterpillar/block/StorageBlock.java
@@ -34,6 +34,7 @@ import java.util.Map;
 import java.util.stream.Stream;
 
 public class StorageBlock extends DrillBaseBlock {
+
     public static final EnumProperty<StoragePart> PART = EnumProperty.create("part", StoragePart.class);
 
     private static final Map<Direction, VoxelShape> SHAPES_LEFT = new EnumMap<>(Direction.class);
@@ -43,23 +44,20 @@ public class StorageBlock extends DrillBaseBlock {
     private static final Map<Direction, VoxelShape> SHAPES_RIGHT = new EnumMap<>(Direction.class);
 
     private static final VoxelShape SHAPE_LEFT = Stream.of(
-            Block.box(1, 10, 1, 16, 14, 15),
-            Block.box(1, 0, 1, 16, 10, 15),
-            Block.box(7.5, 7, 0, 9.5, 11, 1)
+        Block.box(1, 10, 1, 16, 14, 15),
+        Block.box(1, 0, 1, 16, 10, 15),
+        Block.box(7.5, 7, 0, 9.5, 11, 1)
     ).reduce((v1, v2) -> Shapes.join(v1, v2, BooleanOp.OR)).get();
 
     private static final VoxelShape SHAPE_BASE = Stream.of(
-            Block.box(6, 0, 0, 10, 6, 16),
-            Block.box(10, 0, 0, 16, 16, 16),
-            Block.box(0, 0, 0, 6, 16, 16),
-            Block.box(6, 10, 0, 10, 16, 16),
-            Block.box(6, 6, 16, 10, 10, 31)
+        Block.box(0, 0, 0, 16, 16, 16),
+        Block.box(6, 6, 16, 10, 10, 32)
     ).reduce((v1, v2) -> Shapes.join(v1, v2, BooleanOp.OR)).get();
 
     private static final VoxelShape SHAPE_RIGHT = Stream.of(
-            Block.box(0, 10, 1, 15, 14, 15),
-            Block.box(0, 0, 1, 15, 10, 15),
-            Block.box(6.5, 7, 0, 8.5, 11, 1)
+        Block.box(0, 10, 1, 15, 14, 15),
+        Block.box(0, 0, 1, 15, 10, 15),
+        Block.box(6.5, 7, 0, 8.5, 11, 1)
     ).reduce((v1, v2) -> Shapes.join(v1, v2, BooleanOp.OR)).get();
 
     public StorageBlock(Properties properties) {